+++ spec/rspec/core/example_group_spec.rb
@@ -1873,7 +1873,7 @@ module RSpec::Core
       # for users. RSpec internals should not add methods here, though.
       expect(rspec_core_methods.map(&:to_sym)).to contain_exactly(
         :described_class, :subject,
-        :is_expected, :should, :should_not,
+        :is_expected,
         :pending, :skip,
         :setup_mocks_for_rspec, :teardown_mocks_for_rspec, :verify_mocks_for_rspec
       )