+++ Changelog.md
@@ -3,6 +3,9 @@
 Breaking Changes:
 
 * Ruby < 2.3 is no longer supported. (Phil Pirozhkov, #2787)
+* Extract `should` syntax (including the non-monkey-patching one liner). (Phil Pirozhkov, #2803)
+* Remove globally-exposed DSL (example and shared group methods
+  in the root scope and on Module). (Phil Pirozhkov, #2803)
 
 Enhancements: